requirements.txt
streamlit==1.32.0
opencv-python==4.9.0.80
face_recognition==1.3.0
pandas==2.2.1
numpy==1.26.4
openpyxl==3.1.2
Pillow==10.2.0
firebase-admin
